+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Added `Fields.Safe.getOneDeclared` and now `Fields.getOneDeclared` delegates to it.
 - Added `Fields.Safe.getOneDeclaredInHierarchy` and now `Fields.getOneDeclaredInHierarchy` delegates to it.
 - Added missing `Fields.IgnoreAccess.setStatic` overloads to pair with `Fields.IgnoreAccess.getStatic` overloads.
+- Added `Classes.isPresent` to check if a class is present in the class path.
+- Added `Classes.getOne(Class, String)` to return a class which has a name built from the given class and the given prefix.
 
 ---